bugfixed parent, self id the first face

+++ source/facetoplane.h
@@ -20,8 +20,8 @@ public:
 
     QVector3D color;
 
-    int parent;
-    int self;
+    long parent;
+    long self;
 
     QVector2D const& get(QVector3D const &vec);
     QVector3D const& get(QVector2D const &vec);
